pci_cap_init: add 82599 VF quirk
Add check when PCIe capability structure is version 0
and VID and DID is 82599 VF.  In this case, the size
of the PCIe cap structure should be the same as a version 2
cap structure.
Documented in 82599 Errata 35, and is still marked "No Fix".
According to Intel, it's in silicon not fw, and needs a sw workaround.

Acked-by: Michael S. Tsirkin <mst@redhat.com>
Signed-off-by: Donald Dutile <ddutile@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>