pci-assign: Allocate entries for all MSI-X vectors
We still only initialize the number used in the host.  This lets
us do direct access based on MSI-X table offset on write without
needing to translate between physical vector space and initalized
vector space.  It's expected that guests will typically use the
majority of the available vectors, so we're likely not allocating
significantly more entires than are used.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>