qemu-kvm: Fix GSI handling with in-kernel irqchip
Interrupt routing with in-kernel irqchip was always a mess. Now this
increased after latest upstream merge: we started to call ioapic_set_irq
also for kernel-handled interrupts.

Reduce this mess again by using the upstream kvm_piix3_gsi_handler. It
is aware of the PIIX3-specific routing in KVM mode and won't call into
the user space APIC model (as we do not raise IRQs >= ISA_NUM_IRQS with
the PIIX3).

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>