qapi: Add ability to receive QAPI/QMP command that has an array/list as input
The QAPI scripts allow for generating commands that receive parameter
input consisting of a list of custom structs, but the QMP input paramter
checking did not support receiving a qlist as opposed to a qdict for input.

Modified to allow for qlist input parameters, although no argument validation
is performed for commands that accept a list input.