block: VHDX block driver support
This is preliminary and beta support for the VHDX image format,
as specified in:
   "VHDX Format Specification v0.95", published 4/12/2012
    https://www.microsoft.com/en-us/download/details.aspx?id=29681

This RFC patch has the following limitations
    - read-only (no write support yet)
    - does not support differencing files
    - does not make use of iovec or coroutines
    - may likely be broken in novel and interesting ways

I've doen read testing from a dynamic, 127GB VHDX image created
using Hyper-V.  The image was partitioned with a F18 install,
and the partitions could be mounted and files read.

As this is still under active development, and an RFC, you can
assume that any debug print statements will be removed prior
to formal patch submission.