block: Accept node-name arguments for block-commit
This modifies the block operation block-commit so that it will
accept node-name arguments for either 'top' or 'base' BDS.

This means that:

1.) Both "top" and "base" are completely optional.  Previous, "base"
    was optional, but "top" was required.

2.) If "top" is not specified, it defaults to the active layer

3.) "top" and "top-node-name" are mutually exclusive (enforced)

4.) "base" and "base-node-name" are mutually exclusive (enforced)

The preferred and recommended method now of using block-commit
is to specify the BDS to operate on via their node-name arguments.

This also adds an explicit check that base_bs is in the chain of
top_bs, because if they are referenced by their unique node-name
arguments, the discovery process does not intrinsically verify
they are in the same chain.