block: extend block-commit to accept a string for the backing file
On some image chains, QEMU may not always be able to resolve the
filenames properly, when updating the backing file of an image
after a block commit.

For instance, certain relative pathnames may fail, or drives may
have been specified originally by file descriptor (e.g. /dev/fd/???),
or a relative protocol pathname may have been used.

In these instances, QEMU may lack the information to be able to make
the correct choice, but the user or management layer most likely does
have that knowledge.

By extending the block-commit, the user is able to change the backing
file of the overlay image as part of the block-commit operation.

This allows the change to be 'safe', in the sense that if the attempt
to write the backing file image metadata fails, then the block-commit
operation returns failure.

If the commit is from the active layer, then specifying the backing
file string will be treated as an error.

If a backing file string is not specified in the command, the backing
file string to use is determine as it was before.