block: add ability for block-stream to use node-name
This adds the ability for block-stream to use node-name arguments
for base, to specify the backing image to stream from.

Both 'base' and 'base-node-name' are optional, but mutually exclusive.
Either can be specified, but not both together.

The argument for "device" is now optional as well, so long as
base-node-name is specified.  From the node-name, we can determine
the full device chain.