block: add QAPI command to allow live backing file change
This allows a user to change the backing file live, of an open
image.

The image file to modify can be specified 2 ways:

1) 'device' string, and image filename
2) image node-name

Note: this does not cause the backing file itself to be reopened; it
merely changes the backing filename in the image file structure, and
in internal BDS structures.

It is the responsibility of the user to pass a filename string that
can be resolved when the image chain is reopened, and the filename
string is not validated.

A good analogy for this command is that it is a live version of
'qemu-img rebase -u', with respect to change the backing file string.