block: Add overlay BDS pointer into the BlockDriverState struct
Now that node-names can reference an individual BlockDriverState without
needing to navigate downwards from the 'device' level, in order to find
the top-most image (active layer) we need a pointer to the overlay of a
BDS.

This will allow QMP commands to reference an image solely by its
node-name, without also needing to pass in the corresponding 'device'
string.

This function also adds a small helper function, to return the active
layer of any given BDS image.